Allow card-title renaming
This takes advantage of the realization that IDs should NOT have spaces. If there is, the thing before the space should be the identifier and the things afterwards will fulfill that effect.

Rabbi Josh expected this as functionality so we're now providing. For next clients, try communicating this to them instead of letting them experiment and break.